Move agent discovery into proxy.js
Next.js 16 renamed middleware to proxy and the project already had a
proxy.js handling .md rewrites, gone paths, and tracking-param strip.
Merge the Link header and Accept: text/markdown negotiation into it
instead of shipping a parallel middleware.js.
