+++ proxy.js
@@ -1,5 +1,46 @@
 import { NextResponse } from 'next/server'
 
+// Agent discovery resources advertised via RFC 8288 Link headers.
+const LINK_HEADER = [
+  '</.well-known/api-catalog>; rel="api-catalog"; type="application/linkset+json"',
+  '</.well-known/agent-skills/index.json>; rel="https://agentskills.io/rel/index"; type="application/json"',
+  '</llms.txt>; rel="https://llmstxt.org/rel/llms"; type="text/plain"',
+  '</feed.xml>; rel="alternate"; type="application/rss+xml"; title="RSS feed"',
+  '</sitemap.xml>; rel="sitemap"; type="application/xml"',
+  '</robots.txt>; rel="https://www.rfc-editor.org/rfc/rfc9309"; type="text/plain"',
+].join(', ')
+
+function wantsMarkdown(accept) {
+  if (!accept) return false
+  return /(^|,\s*)text\/markdown(\s*;|\s*,|\s*$)/i.test(accept)
+}
+
+function markdownTargetFor(pathname) {
+  const blog = pathname.match(/^\/blog\/([^/]+)\/?$/)
+  if (blog) return `/api/content/${blog[1]}`
+
+  const note = pathname.match(/^\/notes\/([^/]+)\/?$/)
+  if (note) return `/api/content/notes/${note[1]}`
+
+  const collection = pathname.match(/^\/collections\/([^/]+)\/?$/)
+  if (collection) return `/api/content/collections/${collection[1]}`
+
+  if (pathname === '/uses' || pathname === '/uses/') {
+    return '/api/content/pages/uses'
+  }
+  if (pathname === '/about' || pathname === '/about/') {
+    return '/api/content/pages/about'
+  }
+
+  return null
+}
+
+function withDiscoveryHeaders(response) {
+  response.headers.append('Link', LINK_HEADER)
+  response.headers.set('Vary', 'Accept')
+  return response
+}
+
 const STRIP_PARAMS = [
   'URL',
   '_rsc',
@@ -59,6 +100,17 @@ export function proxy(request) {
     return NextResponse.redirect(url, 301)
   }
 
+  // Markdown content negotiation: agents that send `Accept: text/markdown`
+  // get the existing markdown API response in place of the HTML page.
+  if (wantsMarkdown(request.headers.get('accept'))) {
+    const target = markdownTargetFor(pathname)
+    if (target) {
+      const rewriteUrl = request.nextUrl.clone()
+      rewriteUrl.pathname = target
+      return withDiscoveryHeaders(NextResponse.rewrite(rewriteUrl))
+    }
+  }
+
   // Handle .md extension requests
   if (pathname.endsWith('.md')) {
     // Blog posts
@@ -97,7 +149,7 @@ export function proxy(request) {
     }
   }
 
-  return NextResponse.next()
+  return withDiscoveryHeaders(NextResponse.next())
 }
 
 export const config = {
